Add files via upload
The first page of the registration form is "Rform.html".Every connection is made through "Rcheck.php".To store the form I used "Rsave.php"
After submitting the details it redirects to "Rsuccess.html. Then "Validate1.html" and " validate1.php" to check if the email and password exists and then goes to "account.html". Then from there we can edit or delete or logout. For editing I used "select.html,select.php,editing.html and editing.php". For delete it goes to delete.html and delete.php and for logout it goes to logout.html.